fix(console): 总览和定价试算 500——Postgres 的 sum()/avg() 回 numeric，Rust 按 i64/f64 接
用户实拍两屏都是「服务器内部错误，请稍后重试」。后端日志两条不同的解码错：

  error occurred while decoding column 1: Rust type `i64` (as SQL type `INT8`)
    is not compatible with SQL type `NUMERIC`          ← 总览
  error occurred while decoding column 4: Rust type `Option<f64>` (as SQL type
    `FLOAT8`) is not compatible with ...               ← 定价试算

根因同一个：**`sum(bigint)` 在 Postgres 里回的是 `numeric`，不是 `bigint`**（`avg()` 同理）。
不显式转型，sqlx 在解码那一列时直接失败，整个接口 500。

① realtime.rs 收款聚合
   `COALESCE(sum(charged_cents), 0)` → 加 `::bigint`。
   charged_cents 是 bigint，sum 之后成 numeric，而 Rust 声明的是 `Vec<(String, i64)>`。
   **这条是 08-30 17:55 那次提交带来的，今天随我的部署上线 —— 是我把总览打挂的。**

② plan_health.rs 用量分布
   CTE 里 `avg(c) AS per_day`，而 `c = sum(m.cost_cents)` 已经是 numeric，avg 之后还是
   numeric；外层四列（三个 percentile_cont + max(per_day)）都按 `Option<f64>` 接。
   改成 `avg(c)::float8 AS per_day`，一处转型修掉四列。
   这条**在我之前就坏了**（该文件最后改动是 08-28），不是本次部署引入的。

两条都在真库上验过类型（只读 SELECT，零改动）：
   收款：`cny|41600|bigint`（真实收款 ¥416.00，类型对了）
   用量：bigint | double precision | double precision | double precision | bigint
        ——和 Rust 元组逐列对齐

全套 982 通过。

+++ server/src/plan_health.rs
@@ -105,7 +105,7 @@ pub async fn admin_plan_health(
              AND m.free_milli_points_spent = 0 \
            GROUP BY 1, 2 \
          ), pu AS ( \
-           SELECT user_id, count(*)::float8 AS active_days, avg(c) AS per_day, sum(reqs) AS reqs \
+           SELECT user_id, count(*)::float8 AS active_days, avg(c)::float8 AS per_day, sum(reqs) AS reqs \
            FROM d GROUP BY 1 \
          ) \
          SELECT count(*)::bigint, \
